driver_mqtt/requirements.txt
pyserial~=3.5
opencv-python~=4.6
paho-mqtt~=1.6
